Create a RESTful service for the website
Create a REST service that provides the website with content. Using
Node.js and Express, two industry standards, I designed and implement a
web service with a RESTful interface. I needed to modify my previous
project that is the front end/client side of the website to use ajax
calls to communicate with the service. The service should be
supported by a postgresql database.